Web development_ Level-1
I created a responsive HTML page with a heading, paragraph, image, and hyperlink, enhancing it with CSS for a modern look. Additionally, I developed a form with input fields (Name, Email, Password), styled it with CSS, and implemented input validation using required and pattern attributes to ensure proper data entry.